# docs/swagger.yaml
basePath: /
definitions:
...
  models.RefundRequestDto:
    properties:
      card_no:
        type: string
      refund_amount:
        type: integer
      refund_ebonus:
        type: integer
      refund_ecoin:
        type: integer
      refund_meter:
        type: string
      refund_tel:
        type: string
    type: object
...
  models.UsersCreateDto:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      password:
        type: string
      s_name:
        type: string
      status:
        type: string
      user_group_id:
        type: integer
      user_role_id:
        type: integer
      username:
        type: string
    type: object
  models.UsersUpdateDto:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      new_password:
        type: string
      password:
        type: string
      s_name:
        type: string
      status:
        type: string
      user_group_id:
        type: integer
      user_role_id:
        type: integer
      username:
        type: string
    type: object
  utils.StandardErrorResponse:
    properties:
      data:
        description: "null"
      message:
        example: Failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.StandardSuccessResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: This is a POS API with Gin and Swagger
  title: New POS APIls
  version: "1.0"
paths:
...
  /api/prize-counter/jubu-jibi:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search member tel
        in: query
        name: memberTel
        type: string
      - description: Search location
        in: query
        name: location
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get Prize Counter Record Jubu Jibi
      tags:
      - Prize Counter
...
  /api/refund-request:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refund Request Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RefundRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create a new refund request
      tags:
      - Refund Request
  /api/refund-request/{tel}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Telephone number
        in: path
        name: tel
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get refund requests by telephone number
      tags:
      - Refund Request
...
  /api/stamp-house/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search member tel
        in: query
        name: memberTel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get Stamp point List with mobile number and start date
      tags:
      - Stamp House
...
  /api/user:
    post:
      consumes:
      - application/json
      parameters:
      - description: User Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UsersCreateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create a new user
      tags:
      - User
...
  /api/user/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - User
    get:
      consumes:
      - application/json
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get user by id
      tags:
      - User
    put:
      consumes:
      - application/json
      parameters:
      - description: User Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UsersUpdateDto'
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update user by ID
      tags:
      - User
  /api/user/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search text
        in: query
        name: search
        type: string
      - description: Username text
        in: query
        name: username
        type: string
      - description: UserRoleId
        in: query
        name: userRoleId
        type: string
      - description: UserGroupId
        in: query
        name: userGroupId
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get user list with paggination
      tags:
      - User
  /api/users:
    get:
      consumes:
      - application/json
      description: Get all users with Bearer token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get all users
      tags:
      - User
  /api/users/{username}:
    get:
      consumes:
      - application/json
      description: Get users by username with Bearer token
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get users by username
      tags:
      - User
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"